backend_python/requirements.txt
# Spare Parts E-Commerce — Python backend
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.0
Flask-Login>=0.6.3
Flask-CORS>=4.0.0
Werkzeug>=3.0.0
python-dotenv>=1.0.0
openai>=1.12.0
Pillow>=10.0.0
requests>=2.31.0
gunicorn>=22.0.0
pytest>=7.4.0
